Collapse wallet controller into service
