+++ browser/extensions/brave_wallet_unittest.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_wallet/browser/brave_wallet_controller.h"
+#include "brave/components/brave_wallet/browser/brave_wallet_service.h"
 
 #include "brave/common/pref_names.h"
 #include "chrome/browser/profiles/profile_manager.h"
@@ -37,13 +37,13 @@ class BraveWalletUnitTest : public testing::Test {
 };
 
 TEST_F(BraveWalletUnitTest, TestGetRandomNonce) {
-  std::string nonce = BraveWalletController::GetRandomNonce();
-  ASSERT_EQ(nonce.size(), BraveWalletController::kNonceByteLength);
+  std::string nonce = BraveWalletService::GetRandomNonce();
+  ASSERT_EQ(nonce.size(), BraveWalletService::kNonceByteLength);
 }
 
 TEST_F(BraveWalletUnitTest, TestGetRandomSeed) {
-  std::string seed = BraveWalletController::GetRandomSeed();
-  ASSERT_EQ(seed.size(), BraveWalletController::kSeedByteLength);
+  std::string seed = BraveWalletService::GetRandomSeed();
+  ASSERT_EQ(seed.size(), BraveWalletService::kSeedByteLength);
 }
 
 TEST_F(BraveWalletUnitTest, TestGetEthereumRemoteClientSeedFromRootSeed) {
@@ -60,7 +60,7 @@ TEST_F(BraveWalletUnitTest, TestGetEthereumRemoteClientSeedFromRootSeed) {
     52, 76, 223, 24, 183, 138, 244, 72
   };
   std::string derived =
-    BraveWalletController::GetEthereumRemoteClientSeedFromRootSeed(
+    BraveWalletService::GetEthereumRemoteClientSeedFromRootSeed(
         std::string(seed, base::size(seed)));
   ASSERT_EQ(derived, std::string(expected_derived_seed,
       base::size(expected_derived_seed)));
@@ -80,7 +80,7 @@ TEST_F(BraveWalletUnitTest, TestBitGoSeedFromRootSeed) {
     45, 203, 71, 123, 188, 29, 224, 203
   };
   std::string derived =
-    BraveWalletController::GetBitGoSeedFromRootSeed(
+    BraveWalletService::GetBitGoSeedFromRootSeed(
         std::string(seed, base::size(seed)));
   ASSERT_EQ(derived, std::string(expected_derived_seed,
       base::size(expected_derived_seed)));
@@ -111,7 +111,7 @@ TEST_F(BraveWalletUnitTest, TestSealSeed) {
     222, 231, 48, 93, 132, 131, 178, 177
   };
   std::string cipher_seed;
-  ASSERT_TRUE(BraveWalletController::SealSeed(
+  ASSERT_TRUE(BraveWalletService::SealSeed(
      std::string(seed, base::size(seed)), std::string(key, base::size(key)),
      std::string(nonce, base::size(nonce)), &cipher_seed));
   ASSERT_EQ(cipher_seed, std::string(expected_cipher_seed,
@@ -143,7 +143,7 @@ TEST_F(BraveWalletUnitTest, TestOpenSeed) {
     232, 187, 188, 220, 160, 187, 212, 28
   };
   std::string seed;
-  ASSERT_TRUE(BraveWalletController::OpenSeed(
+  ASSERT_TRUE(BraveWalletService::OpenSeed(
        std::string(cipher_seed, base::size(cipher_seed)),
        std::string(key, base::size(key)),
        std::string(nonce, base::size(nonce)), &seed));
@@ -157,7 +157,7 @@ TEST_F(BraveWalletUnitTest, TestLoadFromPrefs) {
 
   std::string cipher_seed;
   std::string nonce;
-  ASSERT_TRUE(BraveWalletController::LoadFromPrefs(
+  ASSERT_TRUE(BraveWalletService::LoadFromPrefs(
       ProfileManager::GetActiveUserProfile(), &cipher_seed, &nonce));
 
   const char expected_nonce[12] = {
@@ -188,7 +188,7 @@ TEST_F(BraveWalletUnitTest, TestSaveToPrefs) {
     6, 128, 179, 64, 55, 160, 219, 8,
     222, 231, 48, 93, 132, 131, 178, 177
   };
-  BraveWalletController::SaveToPrefs(
+  BraveWalletService::SaveToPrefs(
       ProfileManager::GetActiveUserProfile(),
       std::string(cipher_seed, base::size(cipher_seed)),
       std::string(nonce, base::size(nonce)));
