+++ components/brave_wallet/browser/BUILD.gn
@@ -11,8 +11,6 @@ source_set("browser") {
     sources = [
         "brave_wallet_service.h",
         "brave_wallet_service.cc",
-        "brave_wallet_controller.h",
-        "brave_wallet_controller.cc",
         "browser_wallet_delegate.h",
     ]
